pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaggle-lab"
version = "0.1.0"
description = "Laboratory for studying statistical analysis and machine learning with Kaggle datasets."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jupyter>=1.1.1",
    "kagglehub[pandas-datasets]>=0.3.8",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.6.9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
